Move DeconzConfigEntry definition
