+++ homeassistant/components/deconz/button.py
@@ -18,8 +18,9 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+from . import DeconzConfigEntry
 from .entity import DeconzDevice, DeconzSceneMixin
-from .hub import DeconzConfigEntry, DeconzHub
+from .hub import DeconzHub
 
 
 @dataclass(frozen=True, kw_only=True)
