+++ homeassistant/components/deconz/climate.py
@@ -32,9 +32,10 @@ from homeassistant.const import ATTR_TEMPERATURE, UnitOfTemperature
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+from . import DeconzConfigEntry
 from .const import ATTR_LOCKED, ATTR_OFFSET, ATTR_VALVE
 from .entity import DeconzDevice
-from .hub import DeconzConfigEntry, DeconzHub
+from .hub import DeconzHub
 
 DECONZ_FAN_SMART = "smart"
