+++ homeassistant/components/deconz/hub/__init__.py
@@ -1,6 +1,6 @@
 """Internal functionality not part of HA infrastructure."""
 
 from .api import get_deconz_api
-from .hub import DeconzConfigEntry, DeconzHub
+from .hub import DeconzHub
 
-__all__ = ["DeconzConfigEntry", "DeconzHub", "get_deconz_api"]
+__all__ = ["DeconzHub", "get_deconz_api"]
