+++ homeassistant/components/deconz/switch.py
@@ -11,9 +11,9 @@ from homeassistant.components.switch import DOMAIN as SWITCH_DOMAIN, SwitchEntit
 from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+from . import DeconzConfigEntry
 from .const import POWER_PLUGS
 from .entity import DeconzDevice
-from .hub import DeconzConfigEntry
 
 
 async def async_setup_entry(
